+++ CHANGELOG.md
@@ -17,7 +17,7 @@ most recent version is listed first.
 - added `naz` benchmarks and related fixes that came from the benchmark runs: https://github.com/komuw/naz/pull/127
   - when smsc returns `ESME_RMSGQFUL`(message queue is full), `naz` will now call the `throttling` handler.  
   - fixed a potential memory leak bug in `naz.correlater.SimpleCorrelater`
-  - added a configurable `connection_timeout` which is the duration that `naz` will wait, for connection related activities with SMSC, before timing out.
+  - added a configurable `socket_timeout` which is the duration that `naz` will wait, for connection/socket related activities with SMSC, before timing out.
   - `naz` is now able to re-establish connection and re-bind if the connection between it and SMSC is disconnected.